requirements.txt
absl-py==2.0.0
asttokens==2.4.1
astunparse==1.6.3
backcall==0.2.0
blinker==1.6.3
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cvzone==1.6.1
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
easyocr==1.7.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
filterpy==1.4.5
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.41.0
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.0.0
imutils==0.5.4
ipykernel==6.26.0
ipython==8.16.1
# ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.5.0
jupyter_core==5.5.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
numpy==1.25.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.20.0
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
Pygments==2.16.1
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.0.1
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.1
seaborn==0.12.2
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
tifffile==2023.9.26
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.13.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.203
urllib3==2.0.3
wcwidth==0.2.8
Werkzeug==3.0.1
wrapt==1.15.0
zipp==3.16.2
